+++ Game-Bot/MovPrefix.h
@@ -5,10 +5,10 @@ class AssignPrefix {
 public:
 	string NameOfItem = {};
 	uint16_t MinHp = {};
-	uint8_t IdMov = {};
+	uint16_t IdMov = {};
 
 	AssignPrefix();
-	AssignPrefix(string NameOfItem, uint16_t MinHp, uint8_t IdMov);
+	AssignPrefix(string NameOfItem, uint16_t MinHp, uint16_t IdMov);
 
 public:
 	static std::vector<AssignPrefix> assignPrefix;
